+++ backend/models/databases/supabase/__init__.py
@@ -7,6 +7,5 @@ from models.databases.supabase.chats import Chats
 from models.databases.supabase.files import File
 from models.databases.supabase.knowledge import Knowledges
 from models.databases.supabase.notifications import Notifications
-from models.databases.supabase.onboarding import Onboarding
 from models.databases.supabase.user_usage import UserUsage
 from models.databases.supabase.vectors import Vector
