+++ backend/modules/onboarding/controller/onboarding_routes.py
@@ -5,11 +5,12 @@ from middlewares.auth import (  # Assuming you have a get_current_user function
 )
 from modules.onboarding.dto.inputs import OnboardingUpdatableProperties
 from modules.onboarding.entity.onboarding import OnboardingStates
+from modules.onboarding.service.onboarding_service import OnboardingService
 from modules.user.entity.user_identity import UserIdentity
-from repository.onboarding.get_user_onboarding import get_user_onboarding
-from repository.onboarding.update_user_onboarding import update_user_onboarding
 
 onboarding_router = APIRouter()
+
+onboardingService = OnboardingService()
 
 
 @onboarding_router.get(
@@ -24,7 +25,7 @@ async def get_user_onboarding_handler(
     Get user onboarding information for the current user
     """
 
-    return get_user_onboarding(current_user.id)
+    return onboardingService.get_user_onboarding(current_user.id)
 
 
 @onboarding_router.put(
@@ -40,4 +41,4 @@ async def update_user_onboarding_handler(
     Update user onboarding information for the current user
     """
 
-    return update_user_onboarding(current_user.id, onboarding)
+    return onboardingService.update_user_onboarding(current_user.id, onboarding)
