+++ backend/modules/onboarding/service/__init__.py
@@ -1,0 +1,1 @@
+from .onboarding_service import OnboardingService
